+++ appyx-components/internal/test-drive/common/src/androidMain/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.bumble.appyx.components.testdrive" />
+<manifest package="com.bumble.appyx.components.internal.testdrive" />
