+++ appyx-components/internal/test-drive/common/src/commonMain/kotlin/com/bumble/appyx/components/internal/testdrive/ui/simple/TargetUiState.kt
@@ -1,0 +1,13 @@
+package com.bumble.appyx.components.internal.testdrive.ui.simple
+
+import com.bumble.appyx.interactions.core.ui.property.impl.BackgroundColor
+import com.bumble.appyx.interactions.core.ui.property.impl.Position
+import com.bumble.appyx.interactions.core.ui.property.impl.RotationZ
+import com.bumble.appyx.interactions.core.ui.state.MutableUiStateSpecs
+
+@Suppress("unused")
+@MutableUiStateSpecs
+class TargetUiState(
+    val position: Position.Target,
+    val backgroundColor: BackgroundColor.Target,
+)
