--- appyx-components/internal/test-drive/common/src/commonMain/kotlin/com/bumble/appyx/components/testdrive/ui/MaterialColors.kt
@@ -1,298 +1,0 @@
-package com.bumble.appyx.components.testdrive.ui
-
-import androidx.compose.ui.graphics.Color
-
-// reds
-val md_red_50 = Color(0xFFFFEBEE)
-val md_red_100 = Color(0xFFFFCDD2)
-val md_red_200 = Color(0xFFEF9A9A)
-val md_red_300 = Color(0xFFE57373)
-val md_red_400 = Color(0xFFEF5350)
-val md_red_500 = Color(0xFFF44336)
-val md_red_600 = Color(0xFFE53935)
-val md_red_700 = Color(0xFFD32F2F)
-val md_red_800 = Color(0xFFC62828)
-val md_red_900 = Color(0xFFB71C1C)
-val md_red_A100 = Color(0xFFFF8A80)
-val md_red_A200 = Color(0xFFFF5252)
-val md_red_A400 = Color(0xFFFF1744)
-val md_red_A700 = Color(0xFFD50000)
-
-//  pinks
-val md_pink_50 = Color(0xFFFCE4EC)
-val md_pink_100 = Color(0xFFF8BBD0)
-val md_pink_200 = Color(0xFFF48FB1)
-val md_pink_300 = Color(0xFFF06292)
-val md_pink_400 = Color(0xFFEC407A)
-val md_pink_500 = Color(0xFFE91E63)
-val md_pink_600 = Color(0xFFD81B60)
-val md_pink_700 = Color(0xFFC2185B)
-val md_pink_800 = Color(0xFFAD1457)
-val md_pink_900 = Color(0xFF880E4F)
-val md_pink_A100 = Color(0xFFFF80AB)
-val md_pink_A200 = Color(0xFFFF4081)
-val md_pink_A400 = Color(0xFFF50057)
-val md_pink_A700 = Color(0xFFC51162)
-
-//  purples
-val md_purple_50 = Color(0xFFF3E5F5)
-val md_purple_100 = Color(0xFFE1BEE7)
-val md_purple_200 = Color(0xFFCE93D8)
-val md_purple_300 = Color(0xFFBA68C8)
-val md_purple_400 = Color(0xFFAB47BC)
-val md_purple_500 = Color(0xFF9C27B0)
-val md_purple_600 = Color(0xFF8E24AA)
-val md_purple_700 = Color(0xFF7B1FA2)
-val md_purple_800 = Color(0xFF6A1B9A)
-val md_purple_900 = Color(0xFF4A148C)
-val md_purple_A100 = Color(0xFFEA80FC)
-val md_purple_A200 = Color(0xFFE040FB)
-val md_purple_A400 = Color(0xFFD500F9)
-val md_purple_A700 = Color(0xFFAA00FF)
-
-//  deep purples
-val md_deep_purple_50 = Color(0xFFEDE7F6)
-val md_deep_purple_100 = Color(0xFFD1C4E9)
-val md_deep_purple_200 = Color(0xFFB39DDB)
-val md_deep_purple_300 = Color(0xFF9575CD)
-val md_deep_purple_400 = Color(0xFF7E57C2)
-val md_deep_purple_500 = Color(0xFF673AB7)
-val md_deep_purple_600 = Color(0xFF5E35B1)
-val md_deep_purple_700 = Color(0xFF512DA8)
-val md_deep_purple_800 = Color(0xFF4527A0)
-val md_deep_purple_900 = Color(0xFF311B92)
-val md_deep_purple_A100 = Color(0xFFB388FF)
-val md_deep_purple_A200 = Color(0xFF7C4DFF)
-val md_deep_purple_A400 = Color(0xFF651FFF)
-val md_deep_purple_A700 = Color(0xFF6200EA)
-
-//  indigo
-val md_indigo_50 = Color(0xFFE8EAF6)
-val md_indigo_100 = Color(0xFFC5CAE9)
-val md_indigo_200 = Color(0xFF9FA8DA)
-val md_indigo_300 = Color(0xFF7986CB)
-val md_indigo_400 = Color(0xFF5C6BC0)
-val md_indigo_500 = Color(0xFF3F51B5)
-val md_indigo_600 = Color(0xFF3949AB)
-val md_indigo_700 = Color(0xFF303F9F)
-val md_indigo_800 = Color(0xFF283593)
-val md_indigo_900 = Color(0xFF1A237E)
-val md_indigo_A100 = Color(0xFF8C9EFF)
-val md_indigo_A200 = Color(0xFF536DFE)
-val md_indigo_A400 = Color(0xFF3D5AFE)
-val md_indigo_A700 = Color(0xFF304FFE)
-
-// blue
-val md_blue_50 = Color(0xFFE3F2FD)
-val md_blue_100 = Color(0xFFBBDEFB)
-val md_blue_200 = Color(0xFF90CAF9)
-val md_blue_300 = Color(0xFF64B5F6)
-val md_blue_400 = Color(0xFF42A5F5)
-val md_blue_500 = Color(0xFF2196F3)
-val md_blue_600 = Color(0xFF1E88E5)
-val md_blue_700 = Color(0xFF1976D2)
-val md_blue_800 = Color(0xFF1565C0)
-val md_blue_900 = Color(0xFF0D47A1)
-val md_blue_A100 = Color(0xFF82B1FF)
-val md_blue_A200 = Color(0xFF448AFF)
-val md_blue_A400 = Color(0xFF2979FF)
-val md_blue_A700 = Color(0xFF2962FF)
-
-//  light blue
-val md_light_blue_50 = Color(0xFFE1F5FE)
-val md_light_blue_100 = Color(0xFFB3E5FC)
-val md_light_blue_200 = Color(0xFF81D4fA)
-val md_light_blue_300 = Color(0xFF4fC3F7)
-val md_light_blue_400 = Color(0xFF29B6FC)
-val md_light_blue_500 = Color(0xFF03A9F4)
-val md_light_blue_600 = Color(0xFF039BE5)
-val md_light_blue_700 = Color(0xFF0288D1)
-val md_light_blue_800 = Color(0xFF0277BD)
-val md_light_blue_900 = Color(0xFF01579B)
-val md_light_blue_A100 = Color(0xFF80D8FF)
-val md_light_blue_A200 = Color(0xFF40C4FF)
-val md_light_blue_A400 = Color(0xFF00B0FF)
-val md_light_blue_A700 = Color(0xFF0091EA)
-
-//  cyan
-val md_cyan_50 = Color(0xFFE0F7FA)
-val md_cyan_100 = Color(0xFFB2EBF2)
-val md_cyan_200 = Color(0xFF80DEEA)
-val md_cyan_300 = Color(0xFF4DD0E1)
-val md_cyan_400 = Color(0xFF26C6DA)
-val md_cyan_500 = Color(0xFF00BCD4)
-val md_cyan_600 = Color(0xFF00ACC1)
-val md_cyan_700 = Color(0xFF0097A7)
-val md_cyan_800 = Color(0xFF00838F)
-val md_cyan_900 = Color(0xFF006064)
-val md_cyan_A100 = Color(0xFF84FFFF)
-val md_cyan_A200 = Color(0xFF18FFFF)
-val md_cyan_A400 = Color(0xFF00E5FF)
-val md_cyan_A700 = Color(0xFF00B8D4)
-
-//  teal
-val md_teal_50 = Color(0xFFE0F2F1)
-val md_teal_100 = Color(0xFFB2DFDB)
-val md_teal_200 = Color(0xFF80CBC4)
-val md_teal_300 = Color(0xFF4DB6AC)
-val md_teal_400 = Color(0xFF26A69A)
-val md_teal_500 = Color(0xFF009688)
-val md_teal_600 = Color(0xFF00897B)
-val md_teal_700 = Color(0xFF00796B)
-val md_teal_800 = Color(0xFF00695C)
-val md_teal_900 = Color(0xFF004D40)
-val md_teal_A100 = Color(0xFFA7FFEB)
-val md_teal_A200 = Color(0xFF64FFDA)
-val md_teal_A400 = Color(0xFF1DE9B6)
-val md_teal_A700 = Color(0xFF00BFA5)
-
-//  green
-val md_green_50 = Color(0xFFE8F5E9)
-val md_green_100 = Color(0xFFC8E6C9)
-val md_green_200 = Color(0xFFA5D6A7)
-val md_green_300 = Color(0xFF81C784)
-val md_green_400 = Color(0xFF66BB6A)
-val md_green_500 = Color(0xFF4CAF50)
-val md_green_600 = Color(0xFF43A047)
-val md_green_700 = Color(0xFF388E3C)
-val md_green_800 = Color(0xFF2E7D32)
-val md_green_900 = Color(0xFF1B5E20)
-val md_green_A100 = Color(0xFFB9F6CA)
-val md_green_A200 = Color(0xFF69F0AE)
-val md_green_A400 = Color(0xFF00E676)
-val md_green_A700 = Color(0xFF00C853)
-
-// light green
-val md_light_green_50 = Color(0xFFF1F8E9)
-val md_light_green_100 = Color(0xFFDCEDC8)
-val md_light_green_200 = Color(0xFFC5E1A5)
-val md_light_green_300 = Color(0xFFAED581)
-val md_light_green_400 = Color(0xFF9CCC65)
-val md_light_green_500 = Color(0xFF8BC34A)
-val md_light_green_600 = Color(0xFF7CB342)
-val md_light_green_700 = Color(0xFF689F38)
-val md_light_green_800 = Color(0xFF558B2F)
-val md_light_green_900 = Color(0xFF33691E)
-val md_light_green_A100 = Color(0xFFCCFF90)
-val md_light_green_A200 = Color(0xFFB2FF59)
-val md_light_green_A400 = Color(0xFF76FF03)
-val md_light_green_A700 = Color(0xFF64DD17)
-
-//  lime
-val md_lime_50 = Color(0xFFF9FBE7)
-val md_lime_100 = Color(0xFFF0F4C3)
-val md_lime_200 = Color(0xFFE6EE9C)
-val md_lime_300 = Color(0xFFDCE775)
-val md_lime_400 = Color(0xFFD4E157)
-val md_lime_500 = Color(0xFFCDDC39)
-val md_lime_600 = Color(0xFFC0CA33)
-val md_lime_700 = Color(0xFFA4B42B)
-val md_lime_800 = Color(0xFF9E9D24)
-val md_lime_900 = Color(0xFF827717)
-val md_lime_A100 = Color(0xFFF4FF81)
-val md_lime_A200 = Color(0xFFEEFF41)
-val md_lime_A400 = Color(0xFFC6FF00)
-val md_lime_A700 = Color(0xFFAEEA00)
-
-// yellow
-val md_yellow_50 = Color(0xFFFFFDE7)
-val md_yellow_100 = Color(0xFFFFF9C4)
-val md_yellow_200 = Color(0xFFFFF590)
-val md_yellow_300 = Color(0xFFFFF176)
-val md_yellow_400 = Color(0xFFFFEE58)
-val md_yellow_500 = Color(0xFFFFEB3B)
-val md_yellow_600 = Color(0xFFFDD835)
-val md_yellow_700 = Color(0xFFFBC02D)
-val md_yellow_800 = Color(0xFFF9A825)
-val md_yellow_900 = Color(0xFFF57F17)
-val md_yellow_A100 = Color(0xFFFFFF82)
-val md_yellow_A200 = Color(0xFFFFFF00)
-val md_yellow_A400 = Color(0xFFFFEA00)
-val md_yellow_A700 = Color(0xFFFFD600)
-
-// amber
-val md_amber_50 = Color(0xFFFFF8E1)
-val md_amber_100 = Color(0xFFFFECB3)
-val md_amber_200 = Color(0xFFFFE082)
-val md_amber_300 = Color(0xFFFFD54F)
-val md_amber_400 = Color(0xFFFFCA28)
-val md_amber_500 = Color(0xFFFFC107)
-val md_amber_600 = Color(0xFFFFB300)
-val md_amber_700 = Color(0xFFFFA000)
-val md_amber_800 = Color(0xFFFF8F00)
-val md_amber_900 = Color(0xFFFF6F00)
-val md_amber_A100 = Color(0xFFFFE57F)
-val md_amber_A200 = Color(0xFFFFD740)
-val md_amber_A400 = Color(0xFFFFC400)
-val md_amber_A700 = Color(0xFFFFAB00)
-
-// orange
-val md_orange_50 = Color(0xFFFFF3E0)
-val md_orange_100 = Color(0xFFFFE0B2)
-val md_orange_200 = Color(0xFFFFCC80)
-val md_orange_300 = Color(0xFFFFB74D)
-val md_orange_400 = Color(0xFFFFA726)
-val md_orange_500 = Color(0xFFFF9800)
-val md_orange_600 = Color(0xFFFB8C00)
-val md_orange_700 = Color(0xFFF57C00)
-val md_orange_800 = Color(0xFFEF6C00)
-val md_orange_900 = Color(0xFFE65100)
-val md_orange_A100 = Color(0xFFFFD180)
-val md_orange_A200 = Color(0xFFFFAB40)
-val md_orange_A400 = Color(0xFFFF9100)
-val md_orange_A700 = Color(0xFFFF6D00)
-
-// deep orange
-val md_deep_orange_50 = Color(0xFFFBE9A7)
-val md_deep_orange_100 = Color(0xFFFFCCBC)
-val md_deep_orange_200 = Color(0xFFFFAB91)
-val md_deep_orange_300 = Color(0xFFFF8A65)
-val md_deep_orange_400 = Color(0xFFFF7043)
-val md_deep_orange_500 = Color(0xFFFF5722)
-val md_deep_orange_600 = Color(0xFFF4511E)
-val md_deep_orange_700 = Color(0xFFE64A19)
-val md_deep_orange_800 = Color(0xFFD84315)
-val md_deep_orange_900 = Color(0xFFBF360C)
-val md_deep_orange_A100 = Color(0xFFFF9E80)
-val md_deep_orange_A200 = Color(0xFFFF6E40)
-val md_deep_orange_A400 = Color(0xFFFF3D00)
-val md_deep_orange_A700 = Color(0xFFDD2600)
-
-// brown
-val md_brown_50 = Color(0xFFEFEBE9)
-val md_brown_100 = Color(0xFFD7CCC8)
-val md_brown_200 = Color(0xFFBCAAA4)
-val md_brown_300 = Color(0xFFA1887F)
-val md_brown_400 = Color(0xFF8D6E63)
-val md_brown_500 = Color(0xFF795548)
-val md_brown_600 = Color(0xFF6D4C41)
-val md_brown_700 = Color(0xFF5D4037)
-val md_brown_800 = Color(0xFF4E342E)
-val md_brown_900 = Color(0xFF3E2723)
-
-// grey
-val md_grey_50 = Color(0xFFFAFAFA)
-val md_grey_100 = Color(0xFFF5F5F5)
-val md_grey_200 = Color(0xFFEEEEEE)
-val md_grey_300 = Color(0xFFE0E0E0)
-val md_grey_400 = Color(0xFFBDBDBD)
-val md_grey_500 = Color(0xFF9E9E9E)
-val md_grey_600 = Color(0xFF757575)
-val md_grey_700 = Color(0xFF616161)
-val md_grey_800 = Color(0xFF424242)
-val md_grey_900 = Color(0xFF212121)
-val md_black_1000 = Color(0xFF000000)
-val md_white_1000 = Color(0xFFffffff)
-
-// blue grey
-val md_blue_grey_50 = Color(0xFFECEFF1)
-val md_blue_grey_100 = Color(0xFFCFD8DC)
-val md_blue_grey_200 = Color(0xFFB0BBC5)
-val md_blue_grey_300 = Color(0xFF90A4AE)
-val md_blue_grey_400 = Color(0xFF78909C)
-val md_blue_grey_500 = Color(0xFF607D8B)
-val md_blue_grey_600 = Color(0xFF546E7A)
-val md_blue_grey_700 = Color(0xFF455A64)
-val md_blue_grey_800 = Color(0xFF37474F)
-val md_blue_grey_900 = Color(0xFF263238)
-
