--- appyx-components/internal/test-drive/common/src/commonMain/kotlin/com/bumble/appyx/components/testdrive/ui/rotation/TargetUiState.kt
@@ -1,14 +1,0 @@
-package com.bumble.appyx.components.testdrive.ui.rotation
-
-import com.bumble.appyx.interactions.core.ui.property.impl.BackgroundColor
-import com.bumble.appyx.interactions.core.ui.property.impl.Position
-import com.bumble.appyx.interactions.core.ui.property.impl.RotationZ
-import com.bumble.appyx.interactions.core.ui.state.MutableUiStateSpecs
-
-@Suppress("unused")
-@MutableUiStateSpecs
-class TargetUiState(
-    val position: Position.Target,
-    val rotationZ: RotationZ.Target,
-    val backgroundColor: BackgroundColor.Target,
-)
