+++ appyx-components/internal/test-drive/common/src/commonTest/kotlin/com/bumble/appyx/components/internal/testdrive/InteractionTarget.kt
@@ -1,0 +1,5 @@
+package com.bumble.appyx.components.internal.testdrive
+
+enum class InteractionTarget {
+    Child1, Child2, Child3, Child4, Child5, Child6;
+}
