--- appyx-components/internal/test-drive/common/src/commonTest/kotlin/com/bumble/appyx/components/testdrive/InteractionTarget.kt
@@ -1,5 +1,0 @@
-package com.bumble.appyx.components.testdrive
-
-enum class InteractionTarget {
-    Child1, Child2, Child3, Child4, Child5, Child6;
-}
